Unity on the works bug fixed and new features!
after testing this protocol for a while I decided to make it a simple network layer for Unity.
changes:
added a "IsConnected" method for the ClientManager.
added the ability to rejoin a server or close connection and open a new one to a different server using the same ClientManager object.
-added simple Unity support which is still in beta levels (many bugs)
-changed the delegates of the ServerManager events so they will use the same delegate since they work similarly but called on different times (on client connect and on client disconnect)
-fixed small bugs that could be caused due to many threads working at the same time on the same variables (still investigating some bugs)